Introduce DisableFormat that prevents formatting
Summary:
And "none" pseudo-style indicating that formatting should be not applied.

(1) Using .clang-format with "DisableFormat: true" effectively prevents
formatting for all files within the folder containing such .clang-format file.

(2) Using -fallback-style=none together with -style=file prevents formatting
when .clang-format is not found, which can be used in on-save callback.

Reviewers: djasper

Subscribers: cfe-commits, klimek, ono

Differential Revision: http://reviews.llvm.org/D3876